Added functionality to add and delete buttons
Now if the user clicks the add button a row is added to the html table. The values in the row are hard coded. If the user clicks the delete a button, a row is deleted from the table. Next, need to look into how to format the new rows. Then look at HTML entities.