Fixed issues with Add Show button
Changed hover highlight color to yellow. Added maxlength attribute to input elements. Updated component to append new shows so that the styling is identical for the new rows. The delete is no longer working as expected so need to look into this. Blank entries are also not being allowed but no feedback is given to the user.