Add and Delete button fixes
Updated buttons at top of screen to function as links to IMDB/Netflix/HBO.  Moved the delete button outside the form; every time delete was being called so was the submit function. Changed the maxlength to 30 from 10 on the text input fields. Fixed the issue with the user being allowed to delete the table header. Application now prevents the user from entering a blank form my by issuing an alert.