Implementing getting and saving functionality
Added Show component in front end and relevant backend code. Need to clean this up a bit but the next focus is on saving a show to the table. Will need to move code out of the app component and into the show module to do this so I can access the form fields to send to the backend.