Implementing Saving
Removed code from app.component.html into the show component.

JSON not being converted to form by backend so nulls are being saved to the database. Need to figure out how to handle ids. Also need to update page to show the new record upon a successful save or provide an error if there is a validation issue.