Save is now functional
User can now save a new show; however, need to set it up so database doesn't require an id to be passed on an insert query, but rather it is generated automatically. Page refreshes upon a successful  submission; this is fine, but slow. Instead, I could just make a call to grab the showList again. Also need to add validation to ensure user is entering valid input. After that, will look to getting the Delete method working.