User can now delete specific records
Moved the icon button from Menu component to the Show component. Will need to refactor this later as the icons should be part of a separate module; however, for now it allowed for an easier time updating the functionality of the Show component. Added a checkbox button and a select all button to allow for deletes to a specific row, rather than just the most recently added record. Tried to add MapSQLParameterSource to the backend but the parameters were not working with the query so switched back to the existing strategy.