Added Export feature
Added excel report which displays all of the shows in the table. It is built  using Apache POI. The filters are not showing on the report. The report is being sent to the front end and opens in the browser; however, it opens with a random filename and not the one given in the backend.
Also added some navigation which takes the user to a new page when they select a page in the table. Now need to build this page out to include form fields which can allow the user to save more details about the show. Added an icon, which displays all the time. It is a PNG that was built using this website: https://www.canva.com/design