Added proxy config
Following this tutorial to try and get my application deployed: https://dasunpubudu.wordpress.com/2018/04/17/deploying-spring-angular-mast-stack-application-in-heroku/. Reverted some changes from a previous attempt to try and deploy via Heroku. I also need to update the application to handle the case where there are no entries in the DB. Additionally, there is something up with the sorting functionality on the Rating field so need to look into that.